Revise story events and DOM structure, add tests
The goal is to make this more self-contained. <style> elements are inside the
story DOM element, and the actual passage is now in a separate .passage element.
That way, it's possible to remove a story from the DOM wholesale.

Events have been revised to match the namespacing approach that Bootstrap and
Zurb Foundation take.