Fix a bug caused by multiple CSS files
Fix a bug caused by having multiple CSS files in the `src/` directory.

`cssnano` does not support globbing (ex. `src/*.css`). The second CSS
file is actually being interpreted as an `output` argument causing it to
be overwritten, and no output to be written to `stdout`.

To work around this, use `cat` to concatenate all CSS files to a single
stream and pass _that_ to `cssnano`.